:hammer: Restructured objective functions
* The objective functions lives in their own file
    - All remaining files have been updated accordingly
