+++ src/booster/predict.rs
@@ -1,6 +1,6 @@
 use super::booster::ContributionsMethod;
 use crate::{
-    objective::Objective, shapley::predict_contributions_row_shapley, tree::tree::Tree, utils::odds, Matrix,
+    objective_functions::Objective, shapley::predict_contributions_row_shapley, tree::tree::Tree, utils::odds, Matrix,
     PerpetualBooster,
 };
 use rayon::prelude::*;
