+++ src/conformal/cqr.rs
@@ -1,4 +1,4 @@
-use crate::{errors::PerpetualError, objective::Objective, utils::percentiles, Matrix, PerpetualBooster};
+use crate::{errors::PerpetualError, objective_functions::Objective, utils::percentiles, Matrix, PerpetualBooster};
 use std::collections::HashMap;
 
 pub type CalData<'a> = (Matrix<'a, f64>, &'a [f64], &'a [f64]); // (x_flat_data, rows, cols), y, alpha
