+++ src/objective_functions/squared_loss.rs
@@ -1,0 +1,75 @@
+use super::ObjectiveFunction;
+
+use crate::{data::FloatData, metrics::Metric, utils::fast_sum};
+use serde::{Deserialize, Serialize};
+
+#[derive(Default)]
+pub struct SquaredLoss {}
+
+impl ObjectiveFunction for SquaredLoss {
+    #[inline]
+    fn calc_loss(y: &[f64], yhat: &[f64], sample_weight: Option<&[f64]>, _quantile: Option<f64>) -> Vec<f32> {
+        match sample_weight {
+            Some(sample_weight) => y
+                .iter()
+                .zip(yhat)
+                .zip(sample_weight)
+                .map(|((y_, yhat_), w_)| {
+                    let s = *y_ - *yhat_;
+                    (s * s * *w_) as f32
+                })
+                .collect(),
+            None => y
+                .iter()
+                .zip(yhat)
+                .map(|(y_, yhat_)| {
+                    let s = *y_ - *yhat_;
+                    (s * s) as f32
+                })
+                .collect(),
+        }
+    }
+
+    fn calc_init(y: &[f64], sample_weight: Option<&[f64]>, _quantile: Option<f64>) -> f64 {
+        match sample_weight {
+            Some(sample_weight) => {
+                let mut ytot: f64 = 0.;
+                let mut ntot: f64 = 0.;
+                for i in 0..y.len() {
+                    ytot += sample_weight[i] * y[i];
+                    ntot += sample_weight[i];
+                }
+                ytot / ntot
+            }
+            None => fast_sum(y) / y.len() as f64,
+        }
+    }
+
+    #[inline]
+    fn calc_grad_hess(
+        y: &[f64],
+        yhat: &[f64],
+        sample_weight: Option<&[f64]>,
+        _quantile: Option<f64>,
+    ) -> (Vec<f32>, Option<Vec<f32>>) {
+        match sample_weight {
+            Some(sample_weight) => {
+                let (g, h) = y
+                    .iter()
+                    .zip(yhat)
+                    .zip(sample_weight)
+                    .map(|((y_, yhat_), w_)| (((yhat_ - *y_) * *w_) as f32, *w_ as f32))
+                    .unzip();
+                (g, Some(h))
+            }
+            None => (
+                y.iter().zip(yhat).map(|(y_, yhat_)| (yhat_ - *y_) as f32).collect(),
+                None,
+            ),
+        }
+    }
+
+    fn default_metric() -> Metric {
+        Metric::RootMeanSquaredLogError
+    }
+}
